Rename default_zone and fix it's initialisation
default_zone is a little bit too generic, so rename it to
firewalld_default_zone.

Once done, the parameter initialisation must be fixed.
The tasks are using the foo | default() syntax, which is
mostly working, except in the following conditional:

when: defaultzone.stdout != default_zone

It will produce an error if default_zone is not defined,
so set the default value inside defaults/main.yml and
remove all the foo | default() things, which are now useless.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>