tasks/main.yml: Install the python 3 firewalld bindings on Debian buster+
In Debian, the firewalld python bindings are not shipped with the firewalld
package so they have to be installed in a different task.

On buster and newer distributions, there's only python 3 versions of the
binding, so no need to care about python 2 vs 3.
Nevertheless, on buster, it's only in backports, otherwise, there's no
binding available.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>