tasks/main.yml: Fix check mode
In check mode, command module is not run, leading to a failure
in the task setting default zone, since the stdout variable is
not defined.

I could have used "check_mode: no" on the command tasks, but
choose to avoid this, since it may lead to failure in check
mode with firewalld not yet installed.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>